+++ lib/src/configuration/better_player_configuration.dart
@@ -36,6 +36,11 @@ class BetterPlayerConfiguration {
   /// Should the placeholder be shown until play is pressed
   final bool showPlaceholderUntilPlay;
 
+  /// Placeholder position of player stack. If false, then placeholder will be
+  /// displayed on the bottom, so user need to hide it manually. Default is
+  /// true.
+  final bool placeholderOnTop;
+
   /// A widget which is placed between the video and the controls
   final Widget overlay;
 
@@ -96,39 +101,46 @@ class BetterPlayerConfiguration {
   ///play on app resumed). Default value is true.
   final bool handleLifecycle;
 
-  const BetterPlayerConfiguration(
-      {this.aspectRatio,
-      this.autoPlay = false,
-      this.startAt,
-      this.looping = false,
-      this.fullScreenByDefault = false,
-      this.placeholder,
-      this.showPlaceholderUntilPlay = false,
-      this.overlay,
-      this.errorBuilder,
-      this.allowedScreenSleep = true,
-      this.fullScreenAspectRatio,
-      this.deviceOrientationsOnFullScreen = const [
-        DeviceOrientation.landscapeLeft,
-        DeviceOrientation.landscapeRight,
-      ],
-      this.systemOverlaysAfterFullScreen = SystemUiOverlay.values,
-      this.deviceOrientationsAfterFullScreen = const [
-        DeviceOrientation.portraitUp,
-        DeviceOrientation.portraitDown,
-        DeviceOrientation.landscapeLeft,
-        DeviceOrientation.landscapeRight,
-      ],
-      this.routePageBuilder,
-      this.eventListener,
-      this.subtitlesConfiguration = const BetterPlayerSubtitlesConfiguration(),
-      this.controlsConfiguration = const BetterPlayerControlsConfiguration(),
-      this.fit = BoxFit.fill,
-      this.rotation = 0,
-      this.playerVisibilityChangedBehavior,
-      this.translations,
-      this.autoDetectFullscreenDeviceOrientation = false,
-      this.handleLifecycle = true});
+  ///Defines flag which enabled/disabled auto dispose on BetterPlayer dispose.
+  ///Default value is true.
+  final bool autoDispose;
+
+  const BetterPlayerConfiguration({
+    this.aspectRatio,
+    this.autoPlay = false,
+    this.startAt,
+    this.looping = false,
+    this.fullScreenByDefault = false,
+    this.placeholder,
+    this.showPlaceholderUntilPlay = false,
+    this.placeholderOnTop = true,
+    this.overlay,
+    this.errorBuilder,
+    this.allowedScreenSleep = true,
+    this.fullScreenAspectRatio,
+    this.deviceOrientationsOnFullScreen = const [
+      DeviceOrientation.landscapeLeft,
+      DeviceOrientation.landscapeRight,
+    ],
+    this.systemOverlaysAfterFullScreen = SystemUiOverlay.values,
+    this.deviceOrientationsAfterFullScreen = const [
+      DeviceOrientation.portraitUp,
+      DeviceOrientation.portraitDown,
+      DeviceOrientation.landscapeLeft,
+      DeviceOrientation.landscapeRight,
+    ],
+    this.routePageBuilder,
+    this.eventListener,
+    this.subtitlesConfiguration = const BetterPlayerSubtitlesConfiguration(),
+    this.controlsConfiguration = const BetterPlayerControlsConfiguration(),
+    this.fit = BoxFit.fill,
+    this.rotation = 0,
+    this.playerVisibilityChangedBehavior,
+    this.translations,
+    this.autoDetectFullscreenDeviceOrientation = false,
+    this.handleLifecycle = true,
+    this.autoDispose = true,
+  });
 
   BetterPlayerConfiguration copyWith({
     double aspectRatio,
@@ -137,6 +149,8 @@ class BetterPlayerConfiguration {
     bool looping,
     bool fullScreenByDefault,
     Widget placeholder,
+    bool showPlaceholderUntilPlay,
+    bool placeholderOnTop,
     Widget overlay,
     bool showControlsOnInitialize,
     Widget Function(BuildContext context, String errorMessage) errorBuilder,
@@ -162,6 +176,9 @@ class BetterPlayerConfiguration {
       looping: looping ?? this.looping,
       fullScreenByDefault: fullScreenByDefault ?? this.fullScreenByDefault,
       placeholder: placeholder ?? this.placeholder,
+      showPlaceholderUntilPlay:
+          showPlaceholderUntilPlay ?? this.showPlaceholderUntilPlay,
+      placeholderOnTop: placeholderOnTop ?? this.placeholderOnTop,
       overlay: overlay ?? this.overlay,
       errorBuilder: errorBuilder ?? this.errorBuilder,
       allowedScreenSleep: allowedScreenSleep ?? this.allowedScreenSleep,